refactor: split code into modular structure for better organization
